Move load_generators from Rails::Application to Rails::Engine
